python/requirements.txt
pyserial
flask
pymongo